Fix PR #2158 repo lint blockers.
Use the required colonizethis_test import in the new logic test and split new debug flip capital tests into a dedicated app test file to keep scope test file size under repo lint limits.
